+++ src/Plivo/Resource/Meta.cs
@@ -1,4 +1,5 @@
 using System.Threading;
+using Newtonsoft.Json;
 
 namespace Plivo.Resource
 {
@@ -17,25 +18,31 @@ namespace Plivo.Resource
         /// Gets or sets the next.
         /// </summary>
         /// <value>The next.</value>
+        [JsonProperty("next")]
         public string Next { get; set; }
 
         /// <summary>
         /// Gets or sets the offset.
         /// </summary>
         /// <value>The offset.</value>
+        [JsonProperty("offset")]
         public uint Offset { get; set; }
 
         /// <summary>
         /// Gets or sets the previous.
         /// </summary>
         /// <value>The previous.</value>
+        [JsonProperty("previous")]
         public string Previous { get; set; }
 
         /// <summary>
         /// Gets or sets the total count.
         /// </summary>
         /// <value>The total count.</value>
+        [JsonProperty("total_count")]
         public uint TotalCount { get; set; }
+        
+        public uint Count { get; set; }
 
         /// <summary>
         /// Returns a <see cref="T:System.String"/> that represents the current <see cref="T:plivo.Resource.Meta"/>.
@@ -48,7 +55,8 @@ namespace Plivo.Resource
                 "Next: " + Next + "\n" +
                 "Offset: " + Offset + "\n" +
                 "Previous: " + Previous + "\n" +
-                "TotalCount: " + TotalCount + "\n";
+                "TotalCount: " + TotalCount + "\n" +
+                "Count: " + Count + "\n";
         }
     }
 }